requirements.txt
flask
flask-cors
pandas
sqlalchemy
mysql-connector-python
google-ads
python-dotenv
